Publish tag before generating release notes
This change addresses some issues with generating release notes:
- notes were being generated from a ref (branch), there was nothing to
  guarantee this did not change when running the release code
- rpc-differ, and reno when support is added in rpc-openstack, require
  the tag name to produce correctly worded notes

To ensure the tag is created at the start of the process, a new
function `publish_tag` is added to handle pushing the tag to the
repository. This removes that burden from `create_release` with the
added benefit that new tags are annotated (GitHub only creates
lightweight tags).

To enable the tag to be pushed the job now provides access to the
Jenkins GitHub SSH key.

`generate_release_notes` now sets `RE_HOOK_VERSION` to be the version
and not the ref, this is both more intuitive and means that the tag is
now available by the consumer project scripts.