RE-1075 Extend standard jobs to support Docker
To enable the standard jobs to use Docker containers or OpenStack
compute instances, this change adds a new function,
`standard_job_slave`, which manages the slave set-up. The type of slave
a build uses is determined by the parameter `SLAVE_TYPE`, the default
set in the jobs is `"instance"`, to use a Docker container set it to
`"container"`.