RE-1294 Fix job artifact upload to use the RE hook
Currently the job artifact upload assumes the directory
where the artifacts live, rather than making use of the
implemented hook to find out where it is.

This patch corrects the process to use the right hooks,
but also implements a fall back for any jobs which are
not using the hook.